move date-related utils to a separate module
