+++ zyte_common_items/pages/base.py
@@ -1,24 +1,10 @@
-from datetime import datetime
-
-try:
-    from datetime import UTC
-except ImportError:
-    # Python < 3.11
-    from datetime import timezone
-
-    UTC = timezone.utc
-
-
 import attrs
 from web_poet import ItemPage, RequestUrl, WebPage, field
 from web_poet.pages import ItemT
 
-from ..util import format_datetime, metadata_processor
+from .._dateutils import utcnow_formatted
+from ..util import metadata_processor
 from .mixins import HasMetadata, MetadataT
-
-
-def _date_downloaded_now():
-    return format_datetime(datetime.now(UTC))
 
 
 class _BasePage(ItemPage[ItemT], HasMetadata[MetadataT]):
@@ -32,7 +18,7 @@ class _BasePage(ItemPage[ItemT], HasMetadata[MetadataT]):
         value = self.metadata_cls()
         attributes = dir(value)
         if "dateDownloaded" in attributes:
-            value.dateDownloaded = _date_downloaded_now()  # type: ignore
+            value.dateDownloaded = utcnow_formatted()  # type: ignore
         if "probability" in attributes:
             value.probability = 1.0  # type: ignore
         return value
@@ -48,7 +34,7 @@ class _BasePage(ItemPage[ItemT], HasMetadata[MetadataT]):
         metadata = self.metadata_cls()
         metadata_attributes = dir(metadata)
         if "dateDownloaded" in metadata_attributes:
-            metadata.dateDownloaded = _date_downloaded_now()  # type: ignore
+            metadata.dateDownloaded = utcnow_formatted()  # type: ignore
         if "probability" in metadata_attributes:
             metadata.probability = 0.0  # type: ignore
         return self.item_cls(  # type: ignore
